Shooting Mechanic
Got the shooting to work but you can only shoot one time. Need to get it
to work where you will shoot as fast as you can press it. Also, you stop
moving when the space button is pressed. Could fix it by not using the
direction variable to check the shot.